requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
drf-spectacular==0.27.2
drf-yasg==1.21.7
flake8==7.1.0
flake8-isort==6.1.1
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
referencing==0.35.1
rpds-py==0.19.0
setuptools==71.0.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
